Move _process_updates into its own module.
